+++ Mobile Buy SDK/Mobile Buy SDK/Payment Providers/BUYPaymentProvider.h
@@ -26,6 +26,10 @@
 
 @import UIKit;
 
+#if !TARGET_OS_TV
+@import PassKit;
+#endif
+
 #import "BUYClient.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -43,6 +47,7 @@ extern NSString *const BUYPaymentProviderDidCompleteCheckoutNotificationKey;
 
 @required
 
+#if !TARGET_OS_WATCH
 /**
  *  Called when a view controller needs to be presented
  *
@@ -50,6 +55,7 @@ extern NSString *const BUYPaymentProviderDidCompleteCheckoutNotificationKey;
  *  @param controller the `UIViewController` to be presented
  */
 - (void)paymentProvider:(id <BUYPaymentProvider>)provider wantsControllerPresented:(UIViewController *)controller;
+#endif
 
 /**
  *  Called when the view controller
@@ -59,6 +65,20 @@ extern NSString *const BUYPaymentProviderDidCompleteCheckoutNotificationKey;
 - (void)paymentProviderWantsControllerDismissed:(id <BUYPaymentProvider>)provider;
 
 @optional
+
+#if !TARGET_OS_TV
+/**
+ *  Called when a PKPaymentAuthorizationController needs to be presented. If this method is implemented, also
+ *  call `presentWithCompletion:` on the controller and specify completion.
+ *
+ *  PKPaymentAuthorizationController is only available in iOS 10.0+. For older versions, use PKPaymentAuthorizationViewController.
+ *
+ *  @param provider   the `BUYPaymentProvider`
+ *  @param controller the `PKPaymentAuthorizationController` to be presented
+ */
+- (void)paymentProvider:(id <BUYPaymentProvider>)provider wantsPaymentControllerPresented:(PKPaymentAuthorizationController *)controller NS_AVAILABLE_IOS(10_0);
+#endif
+
 
 /**
  *  Called when the checkout process has started