+++ Mobile Buy SDK/Mobile Buy SDK/Utils/BUYModelManager+ApplePay.h
@@ -32,14 +32,16 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface BUYModelManager (ApplePay)
 
+#if !TARGET_OS_WATCH
 /**
  *  Creates a BUYAddress from an ABRecordRef
  *
- *  @param record ABRecordRef to create a BUYAddress from
+ *  @param addressRecord ABRecordRef to create a BUYAddress from
  *
  *  @return The BUYAddress created from an ABRecordRef
  */
 - (BUYAddress *)buyAddressWithABRecord:(ABRecordRef)addressRecord NS_DEPRECATED_IOS(8_0, 9_0, "Use the CNContact backed `buyAddressWithContact:` instead");
+#endif
 
 /**
  *  Creates a BUYAddress from a PKContact